Time: 1 ms (99.85%), Space: 42.3 MB (91.33%) - LeetHub

+++ 3754-concatenate-non-zero-digits-and-multiply-by-sum-i/3754-concatenate-non-zero-digits-and-multiply-by-sum-i.java
@@ -2,7 +2,7 @@ class Solution {
     public long sumAndMultiply(int n) {
         long x = 0;
         long sum = 0;
-        int p = 1;
+        long p = 1;
 
         while(n>0){
            int dig = n%10;
